requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
blinker==1.9.0
certifi==2024.12.14
channels==4.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Django==5.1.4
django-environ==0.12.0
django-plotly-dash==2.4.3
django-prometheus==2.3.1
django-rest-framework==0.1.0
django-sslserver==0.22
djangorestframework==3.15.2
dpd-static-support==0.0.5
dpd_components==0.1.0
et_xmlfile==2.0.0
fastapi==0.115.8
Flask==3.0.3
future==1.0.0
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
nest-asyncio==1.6.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.0.0
plotly==5.24.1
prometheus_client==0.21.1
pydantic==2.10.6
pydantic_core==2.27.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
pyserial==3.5
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
pyzbar==0.1.9
requests==2.32.3
retrying==1.3.4
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.45.3
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
uvicorn==0.34.0
Werkzeug==3.0.6
zebra==0.1.0
zipp==3.21.0
zpl==0.1.11
zplgen==0.0.6
